Move functions from dbgTools to MockUtils
