+++ dbms/src/Debug/MockKVStore/MockUtils.h
@@ -14,9 +14,18 @@
 
 #pragma once
 
+#include <Parsers/IAST_fwd.h>
+#include <RaftStoreProxyFFI/ProxyFFI.h>
+#include <Storages/DeltaMerge/DeltaMergeInterfaces.h>
 #include <Storages/KVStore/MultiRaft/RegionMeta.h>
 #include <Storages/KVStore/MultiRaft/RegionPersister.h>
 #include <Storages/KVStore/TiKVHelpers/TiKVRecordFormat.h>
+
+namespace DB
+{
+class KVStore;
+class TMTContext;
+} // namespace DB
 
 namespace DB::RegionBench
 {
@@ -65,5 +74,76 @@ TiKVValue encodeFullLockCfValue(
     const std::vector<std::string> & async_commit,
     const std::vector<uint64_t> & rollback,
     UInt64 generation = 0);
+
+void setupPutRequest(raft_cmdpb::Request * req, const std::string & cf, const TiKVKey & key, const TiKVValue & value);
+void setupDelRequest(raft_cmdpb::Request *, const std::string &, const TiKVKey &);
+
+void encodeRow(const TiDB::TableInfo & table_info, const std::vector<Field> & fields, WriteBuffer & ss);
+
+void insert(
+    const TiDB::TableInfo & table_info,
+    RegionID region_id,
+    HandleID handle_id,
+    ASTs::const_iterator begin,
+    ASTs::const_iterator end,
+    Context & context,
+    const std::optional<std::tuple<Timestamp, UInt8>> & tso_del = {});
+
+void addRequestsToRaftCmd(
+    raft_cmdpb::RaftCmdRequest & request,
+    const TiKVKey & key,
+    const TiKVValue & value,
+    UInt64 prewrite_ts,
+    UInt64 commit_ts,
+    bool del,
+    const String & pk = "pk");
+
+void concurrentBatchInsert(
+    const TiDB::TableInfo & table_info,
+    Int64 concurrent_num,
+    Int64 flush_num,
+    Int64 batch_num,
+    UInt64 min_strlen,
+    UInt64 max_strlen,
+    Context & context);
+
+void remove(const TiDB::TableInfo & table_info, RegionID region_id, HandleID handle_id, Context & context);
+
+Int64 concurrentRangeOperate(
+    const TiDB::TableInfo & table_info,
+    HandleID start_handle,
+    HandleID end_handle,
+    Context & context,
+    Int64 magic_num,
+    bool del);
+
+Field convertField(const TiDB::ColumnInfo & column_info, const Field & field);
+
+TableID getTableID(
+    Context & context,
+    const std::string & database_name,
+    const std::string & table_name,
+    const std::string & partition_id);
+
+const TiDB::TableInfo & getTableInfo(Context & context, const String & database_name, const String & table_name);
+
+EngineStoreApplyRes applyWriteRaftCmd(
+    KVStore & kvstore,
+    raft_cmdpb::RaftCmdRequest && request,
+    UInt64 region_id,
+    UInt64 index,
+    UInt64 term,
+    TMTContext & tmt,
+    ::DB::DM::WriteResult * write_result_ptr = nullptr);
+
+void handleApplySnapshot(
+    KVStore & kvstore,
+    metapb::Region && region,
+    uint64_t peer_id,
+    SSTViewVec,
+    uint64_t index,
+    uint64_t term,
+    std::optional<uint64_t>,
+    TMTContext & tmt);
 
 } // namespace DB::RegionBench
