+++ dbms/src/Debug/dbgTools.h
@@ -37,78 +37,6 @@ class KVStore;
 class TMTContext;
 } // namespace DB
 
-namespace DB::RegionBench
-{
-
-void encodeRow(const TiDB::TableInfo & table_info, const std::vector<Field> & fields, WriteBuffer & ss);
-
-void insert(
-    const TiDB::TableInfo & table_info,
-    RegionID region_id,
-    HandleID handle_id,
-    ASTs::const_iterator begin,
-    ASTs::const_iterator end,
-    Context & context,
-    const std::optional<std::tuple<Timestamp, UInt8>> & tso_del = {});
-
-void addRequestsToRaftCmd(
-    raft_cmdpb::RaftCmdRequest & request,
-    const TiKVKey & key,
-    const TiKVValue & value,
-    UInt64 prewrite_ts,
-    UInt64 commit_ts,
-    bool del,
-    const String pk = "pk");
-
-void concurrentBatchInsert(
-    const TiDB::TableInfo & table_info,
-    Int64 concurrent_num,
-    Int64 flush_num,
-    Int64 batch_num,
-    UInt64 min_strlen,
-    UInt64 max_strlen,
-    Context & context);
-
-void remove(const TiDB::TableInfo & table_info, RegionID region_id, HandleID handle_id, Context & context);
-
-Int64 concurrentRangeOperate(
-    const TiDB::TableInfo & table_info,
-    HandleID start_handle,
-    HandleID end_handle,
-    Context & context,
-    Int64 magic_num,
-    bool del);
-
-Field convertField(const TiDB::ColumnInfo & column_info, const Field & field);
-
-TableID getTableID(
-    Context & context,
-    const std::string & database_name,
-    const std::string & table_name,
-    const std::string & partition_id);
-
-const TiDB::TableInfo & getTableInfo(Context & context, const String & database_name, const String & table_name);
-
-EngineStoreApplyRes applyWriteRaftCmd(
-    KVStore & kvstore,
-    raft_cmdpb::RaftCmdRequest && request,
-    UInt64 region_id,
-    UInt64 index,
-    UInt64 term,
-    TMTContext & tmt,
-    ::DB::DM::WriteResult * write_result_ptr = nullptr);
-
-void handleApplySnapshot(
-    KVStore & kvstore,
-    metapb::Region && region,
-    uint64_t peer_id,
-    SSTViewVec,
-    uint64_t index,
-    uint64_t term,
-    std::optional<uint64_t>,
-    TMTContext & tmt);
-
-} // namespace DB::RegionBench
 
 namespace DB
 {
